refactor: eliminate provisionTestMode adapter struct in favor of direct helper functions
Replace the provisionTestMode struct and its methods with standalone
helper functions (isVMMode, isAKSMachineMode, setProvisioningError,
clearProvisioningError, getCreateCallCount, popCreationResult,
getSubnetID, setZoneAllocError, setSkuNotAvailable) that check
options.FromContext(ctx).ProvisionMode directly. This removes a layer
of indirection and makes the test code clearer about which provision
mode paths are being exercised.

Also:
- Inline runCommonReuseExistingResourceTests at each call site (100%
  mode-split with zero shared logic)
- Extract shared WellKnownLabelEntry type, wellKnownLabelEntries(), and
  nonSchedulableLabelsMap() to eliminate duplication between AKSScriptless
  and BootstrappingClient test contexts
- Document that CNI label tests are legitimately VM-only (AKS Machine
  handles CNI labels server-side)
- Remove all runCommon/runShared function mode parameters and call-site
  arguments
